% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clust_exp.R
\name{em_clust_exp}
\alias{em_clust_exp}
\title{Mixture modeling of exponentially distributed data.}
\usage{
em_clust_exp(data, nclust, itmax = 10000, tol = 10^-6)
}
\arguments{
\item{data}{An n-length vector. Must not be character.}

\item{nclust}{The number of clusters.}

\item{itmax}{The maximum number of iterations allowed. Defaults to 10000.}

\item{tol}{Tuning parameter for convergence. Defaults to 10^-6.}
}
\value{
A list with four elements \code{it} - the number of iterations, \code{clust_prop}
the estimated mixture proportions, \code{clust_params} the estimated mixture parameters, and
\code{mix_est} a vector of the estimated mixture for each data point.
}
\description{
This function uses the EM algorithm to do clustering of k-mixture components
where each component is \eqn{exponential(\lambda)}.
}
\examples{
\dontshow{c1 <- rexp(100, 1); c2 <- rexp(100, 50); c3 <- rexp(100, 100);
c_tot <- c(c1, c2, c3); rm(c1,c2,c3)}
exp_ex <- em_clust_exp(c_tot, nclust= 3)
}

