% Generated by roxygen2 (4.0.1): do not edit by hand
\name{em_clust_mvn}
\alias{em_clust_mvn}
\title{Clustering for Multivariate Normal data via EM}
\usage{
em_clust_mvn(data, nclust, itmax = 10000, tol = 10^-6)
}
\arguments{
\item{data}{An `n x p` data matrix or data frame.}

\item{nclust}{The number of clusters.}

\item{itmax}{The maximum number of iterations allowed. Defaults to 10000.}

\item{tol}{Tuning parameter for convergence. Defaults to 10^-6.}
}
\value{
A list with four elements \code{it} - the number of iterations, \code{clust_prop}
the estimated mixture proportions, \code{clust_params} a list of the estimated mixture parameters,
and \code{mix_est} a vector of the estimated mixture for each data point.
}
\description{
This function uses the EM algorithm to do clustering in P-dimensions.
It assumes all clusters are spherically \eqn{N(\mu_m, \Sigma_m I)}.
}

