% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_pois.R
\name{em_clust_pois}
\alias{em_clust_pois}
\title{Mixture modeling of poisson distributed data.}
\usage{
em_clust_pois(data, nclust, itmax = 10000, tol = 10^-8)
}
\arguments{
\item{data}{An n-length vector. Must not be character.}

\item{nclust}{The number of clusters.}

\item{itmax}{The maximum number of iterations allowed. Defaults to 10000.}

\item{tol}{Tuning parameter for convergence. Defaults to 10^-8.}
}
\value{
A list containing: \code{it} the number of iterations; \code{clust_prop}
the estimated mixture proportions; \code{clust_params} the estimated mixture parameters; 
\code{mix_est} a vector of the estimated mixture for each data point; \code{log_lik} the 
log likelihood of the data; \code{bic} the modeled BIC.
}
\description{
This function uses the EM algorithm to do clustering of k-mixture components
where each component is \eqn{pois(\lambda)}.
}
\examples{
# generate test data
c1 <- rpois(100, 3); c2 <- rpois(100, 20); c3 <- rpois(100, 100)
c_tot <- c(c1, c2, c3); rm(c1,c2,c3)
# run example
pois <- em_clust_pois(c_tot, nclust= 3) 
}
