% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{gen_clust}
\alias{gen_clust}
\title{Random multivariate normal cluster generation}
\usage{
gen_clust(obs, p, mean, sd)
}
\arguments{
\item{obs}{The number of observations you wish to generate.}

\item{p}{The number of dimensions for the multivariate normal cluster to be generated.}

\item{mean}{A vector of means, one for each dimension of the cluster.}

\item{sd}{A vector of standard deviations, one for each dimension of the cluster.}
}
\value{
Returns an `obs x p` matrix where each column $p_i$ is an obs-length Gaussian
distributed vector with \eqn{N(\mu= mean[i], \sigma= sd[i])}.
}
\description{
This function generates a random p-dimensional multivariate normal cluster.
}
\examples{
c1 <- gen_clust(100, 10, mean= c(seq(-8, 10, 2)), sd= rep(1, 10))
}

