% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clust_mvn.R
\name{em_clust_mvn}
\alias{em_clust_mvn}
\title{Clustering for Multivariate Normal data via EM}
\usage{
em_clust_mvn(data, nclust, itmax = 10000, tol = 10^-6)
}
\arguments{
\item{data}{An `n x p` data matrix or data frame.}

\item{nclust}{The number of clusters.}

\item{itmax}{The maximum number of iterations allowed. Defaults to 10000.}

\item{tol}{Tuning parameter for convergence. Defaults to 10^-6.}
}
\value{
A list containing: \code{it} the number of iterations; \code{clust_prop}
the estimated mixture proportions; \code{clust_params} the estimated mixture parameters;
\code{mix_est} a vector of the estimated mixture for each data point; \code{log_lik} the
log likelihood of the data; \code{bic} the modeled BIC.
}
\description{
This function uses the EM algorithm to do clustering in P-dimensions.
It assumes all clusters are spherically \eqn{N(\mu_m, \Sigma_m I)}.
}
\examples{
\dontshow{c1 <- gen_clust(100, 10, mean= c(seq(-8, 10, 2)), sd= rep(1, 10))
c2 <- gen_clust(100, 10, mean= rep(0, 10), sd= rep(2, 10))
c3 <- gen_clust(100, 10, mean= rep(10, 10), sd= rep(1, 10))
c_tot <- rbind(c1,c2,c3); rm(c1,c2,c3)}
mvn <- em_clust_mvn(c_tot, nclust= 3)
}
\seealso{
\code{\link{em_clust_norm}}, \code{\link{em_clust_mvn_miss}}, \code{\link{gen_clust}}
}

