% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clust_mvn_miss.R
\name{em_clust_mvn_miss}
\alias{em_clust_mvn_miss}
\title{Clustering for Multivariate Normal with missing data via EM}
\usage{
em_clust_mvn_miss(data, nclust, itmax = 10000, tol = 10^-8)
}
\arguments{
\item{data}{An `n x p` data matrix or data frame.}

\item{nclust}{The number of clusters.}

\item{itmax}{The maximum number of iterations allowed. Defaults to 10000.}

\item{tol}{Tuning parameter for convergence. Defaults to 10^-8.}
}
\value{
A list containing: \code{it} the number of iterations; \code{clust_prop}
the estimated mixture proportions; \code{clust_params} the estimated mixture parameters;
\code{mix_est} a vector of the estimated mixture for each data point; \code{pseudo_log_lik}
the pseudo log likelihood of the data; \code{bic} the modeled BIC.
and \code{mix_est} a vector of the estimated mixture for each data point.
}
\description{
This function uses the EM algorithm to do clustering in P-dimensions.
It assumes all clusters are spherically \eqn{N(\mu_m, \Sigma_m I)} and allows for
missing data. Observations may have missing data elements but entire rows may
not be missing.
}
\examples{
# generate test data
c1 <- gen_clust(100, 10, mean= c(seq(-8, 10, 2)), sd= rep(1, 10))
c2 <- gen_clust(100, 10, mean= rep(0, 10), sd= rep(2, 10))
c3 <- gen_clust(100, 10, mean= rep(10, 10), sd= rep(1, 10))
c_tot <- rbind(c1,c2,c3); rm(c1,c2,c3)
c_tot <- apply(c_tot, 2, function(x) {
  samp <- sample(1:length(x), floor(length(x) * .2), replace=FALSE)
  x[samp] <- NA
  return(x)
})
# run example
mvn_miss <- em_clust_mvn_miss(c_tot, nclust= 3)
}
\seealso{
\code{\link{em_clust_mvn}}, \code{\link{em_clust_norm}}, \code{\link{gen_clust}}
}

