% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clust_norm.R
\name{em_clust_norm}
\alias{em_clust_norm}
\title{Mixture modeling of normally distributed univariate data.}
\usage{
em_clust_norm(data, nclust, itmax = 10000, tol = 10^-6)
}
\arguments{
\item{data}{An n-length vector. Must not be character.}

\item{nclust}{The number of clusters.}

\item{itmax}{The maximum number of iterations allowed. Defaults to 10000.}

\item{tol}{Tuning parameter for convergence. Defaults to 10^-6.}
}
\value{
A list containing: \code{it} the number of iterations; \code{clust_prop}
the estimated mixture proportions; \code{clust_params} the estimated mixture parameters;
\code{mix_est} a vector of the estimated mixture for each data point; \code{log_lik} the
log likelihood of the data; \code{bic} the modeled BIC.
}
\description{
This function uses the EM algorithm to do clustering of k-mixture components
where each component is one-dimensional \eqn{N(\mu, \sigma^2)}.
}
\examples{
# generate test data
c1 <- rnorm(100, 5, 1); c2 <- rnorm(100, 15, 1); c3 <- rnorm(100, 20, 1)
c_tot <- c(c1, c2, c3); rm(c1,c2,c3)
# run example
norm_ex <- em_clust_norm(c_tot, nclust= 3)
}
\seealso{
\code{\link{em_clust_mvn}}, \code{\link{em_clust_mvn_miss}}, \code{\link{gen_clust}}
}

